Added a script
Added a script for implementing the functions. As can be seen in the comments, the domain may be partitioned to account for changes in the element properties. With the global assembler function (which I am in the process of rewriting), there is no need for user-input. Everything may be passed by parameter through it. Please don't gum up the code by asking the user to enter stuff. Thanks.